working on zbuf
