+++ larp.cpp
@@ -14,7 +14,7 @@
 // Globals
 SDL_Window *g_window = NULL;        // The window we'll be rendering to
 SDL_Renderer *g_renderer = NULL;    // The window renderer
-SDL_Surface *g_zbuffer = NULL;
+SDL_Surface *g_buffer = NULL;       // Surface used as Z-buffer
 
 // Scene
 Model g_model;
@@ -23,47 +23,43 @@ Camera g_camera;
 
 bool init(void)
 {
-    bool success = true;
-
 	// Initialize SDL
-	if(SDL_Init(SDL_INIT_VIDEO) < 0)
-    {
+	if (SDL_Init(SDL_INIT_VIDEO) < 0) {
 		printf("SDL could not initialize! SDL_Error: %s\n", SDL_GetError());
-        success = false;
+        return false;
 	}
-	else
-    {
-        // Set texture filtering to linear
-        if (!SDL_SetHint(SDL_HINT_RENDER_SCALE_QUALITY, "1"))
-        {
-            printf("Warning: Linear texture filtering not enabled");
-        }
 
-		// Create window
-		g_window = SDL_CreateWindow(WINDOW_NAME, SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, SCREEN_WIDTH, SCREEN_HEIGHT, SDL_WINDOW_SHOWN);
-		if(g_window == NULL)
-        {
-			printf("Window could not be created! SDL_Error: %s\n", SDL_GetError());
-            success = false;
-		}
-		else 
-        {
-            // Create renderer for window
-            g_renderer = SDL_CreateRenderer(g_window, -1, SDL_RENDERER_ACCELERATED);
-            if (g_renderer == NULL)
-            {
-                printf("Renderer could not be created. SDL Error: %s\n", SDL_GetError());
-                success = false;
-            }
-            else
-            {
-                // Initialize renderer color
-                SDL_SetRenderDrawColor(g_renderer, 0xFF, 0xFF, 0xFF, 0xFF);
-            }
-        }
+    // Set texture filtering to linear
+    if (!SDL_SetHint(SDL_HINT_RENDER_SCALE_QUALITY, "1")) {
+        printf("Warning: Linear texture filtering not enabled");
     }
 
-    return success;
+    // Create window
+    g_window = SDL_CreateWindow(WINDOW_NAME, SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, SCREEN_WIDTH, SCREEN_HEIGHT, SDL_WINDOW_SHOWN);
+    if(g_window == NULL) {
+        printf("Window could not be created! SDL_Error: %s\n", SDL_GetError());
+        return false;
+    }
+
+    // Create renderer for window
+    g_renderer = SDL_CreateRenderer(g_window, -1, SDL_RENDERER_ACCELERATED);
+    if (g_renderer == NULL) {
+        printf("Renderer could not be created. SDL Error: %s\n", SDL_GetError());
+        return false;
+    }
+    
+    // Initialize renderer color
+    SDL_SetRenderDrawColor(g_renderer, 0xFF, 0xFF, 0xFF, 0xFF);
+
+
+    // Initialize surface
+    g_buffer = SDL_CreateRGBSurfaceWithFormat(0, SCREEN_WIDTH, SCREEN_HEIGHT, 32, SDL_PIXELFORMAT_RGBA8888);
+    if (g_buffer == NULL) {
+        printf("Surface could not be created. SDL Error: %s\n", SDL_GetError());
+        return false;
+    }
+
+    return true;
 }
 
 void end(void) 
@@ -71,8 +67,10 @@ void end(void)
 	// Destroy window
     SDL_DestroyRenderer(g_renderer);
 	SDL_DestroyWindow(g_window);
+	SDL_FreeSurface(g_buffer);
     g_window = NULL;
     g_renderer = NULL;
+    g_buffer = NULL;
 
 	// Quit SDL subsystems
 	SDL_Quit();
@@ -83,10 +81,6 @@ void initScene()
     // Load objects
     g_model = Model();
     g_model.LoadModel(MODEL_PATH);
-
-    // Reset depth buffer
-    // memset(g_zbuffer, 0, sizeof(g_zbuffer));
-
 }
 
 void renderScene()
@@ -95,13 +89,14 @@ void renderScene()
     SDL_SetRenderDrawColor(g_renderer, 0xFF, 0xFF, 0xFF, 0xFF);
     SDL_RenderClear(g_renderer);
 
-    // Reset depth buffer
-    // memset(g_zbuffer, 0, sizeof(g_zbuffer));
+    // Clear the z buffer 
+    // Init alpha (z) channel to max value
+    SDL_FillRect(g_buffer, NULL, SDL_MapRGBA(g_buffer->format, 0, 0, 0, 255));
 
     // Redraw models
-    SDL_SetRenderDrawColor(g_renderer, 0x00, 0x00, 0xFF, 0xFF);
+    // SDL_SetRenderDrawColor(g_renderer, 0x00, 0x00, 0xFF, 0xFF);
     // g_model.DrawEdges(g_camera, g_renderer);
-    // g_model.DrawFaces(g_camera, g_renderer, g_zbuffer);
+    g_model.DrawFaces(g_camera, g_renderer, g_buffer);
 
     //Update screen
     SDL_RenderPresent(g_renderer);
