+++ utils/constants.h
@@ -13,9 +13,9 @@
 // #define MODEL_PATH "assets/dfiles/jcolumn.d"
 // #define MODEL_PATH "assets/dfiles/jbase.d"
 // #define MODEL_PATH "assets/dfiles/floor.d"
-#define MODEL_PATH "assets/dfiles/cow.d"
+// #define MODEL_PATH "assets/dfiles/cow.d"
 // #define MODEL_PATH "assets/dfiles/atc.d" // too large a model
-// #define MODEL_PATH "assets/dfiles/better-ball.d" //very good test
+#define MODEL_PATH "assets/dfiles/better-ball.d" //very good test
 // #define MODEL_PATH "assets/dfiles/house.d"
 // #define MODEL_PATH "assets/dfiles/nteapot6.d"
 // #define MODEL_PATH "assets/dfiles/ronny.d"
